+++ packages/svelte-query/tests/context/context.svelte.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, test } from 'vitest'
 import { render } from '@testing-library/svelte'
 import { getIsRestoringContext } from '../../src/index.js'
-import BaseExample from './BaseExample.svelte'
+import Base from './Base.svelte'
 
 describe('getQueryClientContext', () => {
   test('should throw when called without a client in context', () => {
-    expect(() => render(BaseExample)).toThrow(
+    expect(() => render(Base)).toThrow(
       'No QueryClient was found in Svelte context. Did you forget to wrap your component with QueryClientProvider?',
     )
   })