+++ packages/svelte-query/tests/useIsFetching/useIsFetching.svelte.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 import { fireEvent, render } from '@testing-library/svelte'
 import { QueryClient } from '@tanstack/query-core'
-import BaseExample from './BaseExample.svelte'
+import Base from './Base.svelte'
 
 describe('useIsFetching', () => {
   let queryClient: QueryClient
@@ -17,7 +17,7 @@ describe('useIsFetching', () => {
   })
 
   test('should update as queries start and stop fetching', async () => {
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: { queryClient },
     })