+++ packages/angular-query-experimental/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/angular-query-experimental
 
+## 5.99.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.99.1
+
 ## 5.99.0
 
 ### Patch Changes