+++ packages/react-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/react-query-persist-client
 
+## 5.99.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.99.1
+  - @tanstack/react-query@5.99.1
+
 ## 5.99.0
 
 ### Patch Changes